requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.7.0
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
bcrypt==3.2.2
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.0
colorama==0.4.6
courlan==1.3.2
dateparser==1.2.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.6
faust-cchardet==2.1.19
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.73.1
grpcio-status==1.71.0
h11==0.16.0
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jose==1.0.0
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
passlib==1.7.4
primp==0.15.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
py3langid==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycurl==7.45.6
pydantic==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyee==13.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tld==0.13.1
tomli==2.2.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.0
zstandard==0.23.0
